+++ gui/state_registry.go
@@ -236,6 +236,7 @@ const (
 	nsSvgCache            = "gui.svg_cache"
 	nsSvgDimCache         = "gui.svg_dim_cache"
 	nsSvgAnimSeen         = "gui.svg_anim_seen"
+	nsSvgErrWarned        = "gui.svg.err_warned"
 	nsOpticalOffset       = "gui.optical_offset"
 	nsDragReorder         = "gui.drag_reorder"
 	nsDragReorderIDsMeta  = "gui.drag_reorder.ids_meta"